requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
argcomplete==3.5.1
attrs==24.2.0
black==24.10.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
datamodel-code-generator==0.26.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
genson==1.3.0
gevent==24.11.1
geventhttpclient==2.3.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflect==5.6.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
locust==2.33.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
msgpack==1.1.0
mypy-extensions==1.0.0
narwhals==1.14.1
numpy==2.1.2
openai==1.77.0
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.28.3
psutil==7.0.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
setuptools==76.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
